teclado y pantalla de la chip8 funcionan

+++ cliente/include/chip8/chip8_input.hpp
@@ -5,6 +5,6 @@
 #include <stdbool.h>
 #include <SDL2/SDL.h>
 
-bool chip8inputCapturarTeclado(uint16_t *teclado);
+bool chip8inputCapturarTeclado(uint8_t *teclado);
 
 #endif // !INPUT_H
